updated PayPal routing and templating
