+++ src/Controller/Paypal/AbstractController.php
@@ -1,0 +1,35 @@
+<?php
+
+namespace App\Controller\Paypal;
+
+use App\Service\PaypalService;
+use App\Service\SessionService;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController as SymfonyAbstractController;
+
+/**
+ * Class AbstractController
+ * @package App\Controller\PayPal
+ */
+class AbstractController extends SymfonyAbstractController
+{
+    /**
+     * @var SessionService
+     */
+    protected $sessionService;
+
+    /**
+     * @var PaypalService
+     */
+    protected $paypalService;
+
+    /**
+     * DefaultController constructor.
+     * @param SessionService $sessionService
+     * @param PaypalService $paypalService
+     */
+    public function __construct(SessionService $sessionService, PaypalService $paypalService)
+    {
+        $this->sessionService = $sessionService;
+        $this->paypalService = $paypalService;
+    }
+}
