+++ src/Controller/Paypal/DefaultController.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace App\Controller\Paypal;
+
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpFoundation\RedirectResponse;
+
+/**
+ * Class DefaultController
+ * @package App\Controller\Paypal
+ *
+ * @Route("/paypal", name="paypal-")
+ */
+class DefaultController extends AbstractController
+{
+    /**
+     * @Route("/", name="index")
+     *
+     * @return RedirectResponse|Response
+     */
+    public function index()
+    {
+        $request = Request::createFromGlobals();
+        return $this->redirectToRoute('paypal-anonymous-index', $request->query->all());
+    }
+}
