+++ templates/paypal/anonymous/index.html.twig
@@ -29,7 +29,7 @@
                 "buttonShape":"pill",
                 "buttonSize":"lg",
                 "fullPage":"true",
-                "returnurl":"{{ absolute_url(path('paypal-anonymous-home')) }}"
+                "returnurl":"{{ absolute_url(path('paypal-anonymous-index')) }}"
             });
         });
     </script>
