+++ templates/paypal/anonymous/user-info.html.twig
@@ -31,7 +31,7 @@
         </div>
         <div class="row">
             <div class="col text-center">
-                <form action="{{ path('paypal-account') }}" method="post">
+                <form action="{{ path('paypal-authenticated-index') }}" method="post">
                     <input type="hidden" id="refresh_token" name="refresh_token" value="{{ refresh_token }}">
                     <button type="submit" class="btn btn-primary btn-lg btn-block">Go to MyAccount</button>
                 </form>
