+++ templates/paypal/authenticated/base.html.twig
@@ -7,11 +7,11 @@
     <div class="navbar-brand collapse navbar-collapse" id="navbarSupportedContent">
         {% include "navbar/dropdown.html.twig" %}
         <div class="navbar-nav col justify-content-center">
-            <a class="nav-link text-light" href="{{ path('paypal-account') }}">My Account</a>
-            <a class="nav-link text-light" href="{{ path('paypal-payments') }}">Send Payments</a>
-            <a class="nav-link text-light" href="{{ path('paypal-payouts') }}">Receive Payments</a>
-            <a class="nav-link text-light" href="{{ path('paypal-invoices') }}">Invoices QR</a>
-            <a class="nav-link text-light" href="{{ path('paypal-subscriptions') }}">Subscriptions</a>
+            <a class="nav-link text-light" href="{{ path('paypal-authenticated-index') }}">My Account</a>
+            <a class="nav-link text-light" href="{{ path('paypal-authenticated-payments') }}">Send Payments</a>
+            <a class="nav-link text-light" href="{{ path('paypal-authenticated-payouts') }}">Receive Payments</a>
+            <a class="nav-link text-light" href="{{ path('paypal-authenticated-invoices') }}">Invoices QR</a>
+            <a class="nav-link text-light" href="{{ path('paypal-authenticated-subscriptions') }}">Subscriptions</a>
         </div>
         <ul class="navbar-nav col justify-content-end">
             {% include "paypal/modals/sandbox-credentials.html.twig" %}
@@ -24,7 +24,7 @@
     </div>
     <div class="col text-center">
         <button type="button" class="btn btn-success" onclick="window.location.reload();">Refresh</button>
-        <button type="button" class="btn btn-secondary" onclick="location.href = '{{ path('paypal-logout') }}';">Log Out</button>
+        <button type="button" class="btn btn-secondary" onclick="location.href = '{{ path('paypal-session-logout') }}';">Log Out</button>
     </div>
 </div>
 <div class="container mt-4">
