+++ templates/paypal/modals/sandbox-credentials.html.twig
@@ -1,7 +1,7 @@
 <div class="modal fade" id="credentialsModal" tabindex="-1" role="dialog" aria-labelledby="credentialsModal" aria-hidden="true">
     <div class="modal-dialog" role="document">
         <div class="modal-content">
-            <form action="{{ path('paypal-sandbox-credentials') }}" method="post">
+            <form action="{{ path('paypal-session-sandbox-credentials') }}" method="post">
             <div class="modal-header text-center">
                 <h5 class="modal-title w-100 font-weight-bold">Update SandBox Credentials</h5>
             </div>
@@ -20,14 +20,14 @@
                 </div>
             </div>
             <div class="modal-footer d-flex justify-content-center">
-                <button type="submit" class="btn btn-primary" onclick="location.href = '{{ path('paypal-sandbox-credentials') }}';">Update Now</button>
-                <button type="button" class="btn btn-secondary" onclick="location.href = '{{ path('paypal-logout') }}';">Clear</button>
+                <button type="submit" class="btn btn-primary">Update Now</button>
+                <button type="button" class="btn btn-secondary" onclick="location.href = '{{ path('paypal-session-logout') }}';">Clear</button>
             </div>
             </form>
             <div class="modal-body text-center">
                 <h6><strong>This credentials are saved for 8 hours</strong></h6>
                 <h6>Add the Return URL to SandBox App Settings</h6>
-                <h6 class="font-weight-light">{{ absolute_url(path('paypal-anonymous-home')) }}</h6>
+                <h6 class="font-weight-light">{{ absolute_url(path('paypal-anonymous-index')) }}</h6>
             </div>
         </div>
     </div>
